+++ filament/backend/src/vulkan/VulkanBuffer.cpp
@@ -69,52 +69,38 @@ void VulkanBuffer::loadFromCpu(VulkanContext& context, VulkanStagePool& stagePoo
 
     // Firstly, ensure that the copy finishes before the next draw call.
     // Secondly, in case the user decides to upload another chunk (without ever using the first one)
-    // we need to ensure that this upload completes first.
-
+    // we need to ensure that this upload completes first (hence
+    // dstStageMask=VK_PIPELINE_STAGE_TRANSFER_BIT).
+    VkAccessFlags dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
+    VkPipelineStageFlags dstStageMask = VK_PIPELINE_STAGE_TRANSFER_BIT;
     if (mUsage & VK_BUFFER_USAGE_VERTEX_BUFFER_BIT) {
-
-        VkBufferMemoryBarrier barrier{
-                .sType = VK_STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER,
-                .srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT,
-                .dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT |
-                                 VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT | VK_ACCESS_INDEX_READ_BIT,
-                .srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED,
-                .dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED,
-                .buffer = mGpuBuffer,
-                .size = VK_WHOLE_SIZE
-        };
-
-        vkCmdPipelineBarrier(cmdbuffer,
-                VK_PIPELINE_STAGE_TRANSFER_BIT,
-                VK_PIPELINE_STAGE_TRANSFER_BIT | VK_PIPELINE_STAGE_VERTEX_INPUT_BIT,
-                0, 0, nullptr, 1, &barrier, 0, nullptr);
-
+	dstAccessMask |= VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT;
+	dstStageMask |= VK_PIPELINE_STAGE_VERTEX_INPUT_BIT;
+    } else if (mUsage & VK_BUFFER_USAGE_INDEX_BUFFER_BIT) {
+	dstAccessMask |= VK_ACCESS_INDEX_READ_BIT;
+	dstStageMask |= VK_PIPELINE_STAGE_VERTEX_INPUT_BIT;
+    } else if (mUsage & VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT) {
+	dstAccessMask |= VK_ACCESS_UNIFORM_READ_BIT;
+	// NOTE: ideally dstStageMask would include VERTEX_SHADER_BIT | FRAGMENT_SHADER_BIT, but
+	// this seems to be insufficient on Mali devices. To work around this we are using a more
+	// aggressive ALL_GRAPHICS_BIT barrier.
+	dstStageMask |= VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT;
+    } else if (mUsage & VK_BUFFER_USAGE_STORAGE_BUFFER_BIT) {
+	// TODO: implement me
     }
 
-    if (mUsage & VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT) {
-        // NOTE: ideally dstStageMask would include VERTEX_SHADER_BIT | FRAGMENT_SHADER_BIT, but this
-        // seems to be insufficient on Mali devices. To work around this we are using a more
-        // aggressive ALL_GRAPHICS_BIT barrier.
+    VkBufferMemoryBarrier barrier{
+	    .sType = VK_STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER,
+	    .srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT,
+	    .dstAccessMask = dstAccessMask,
+	    .srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED,
+	    .dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED,
+	    .buffer = mGpuBuffer,
+	    .size = VK_WHOLE_SIZE,
+    };
 
-        VkBufferMemoryBarrier barrier{
-                .sType = VK_STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER,
-                .srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT,
-                .dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT | VK_ACCESS_UNIFORM_READ_BIT,
-                .srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED,
-                .dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED,
-                .buffer = mGpuBuffer,
-                .size = VK_WHOLE_SIZE
-        };
-
-        vkCmdPipelineBarrier(cmdbuffer,
-                VK_PIPELINE_STAGE_TRANSFER_BIT,
-                VK_PIPELINE_STAGE_TRANSFER_BIT | VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT,
-                0, 0, nullptr, 1, &barrier, 0, nullptr);
-    }
-
-    if (mUsage & VK_BUFFER_USAGE_STORAGE_BUFFER_BIT) {
-        // TODO: implement me
-    }
+    vkCmdPipelineBarrier(cmdbuffer, VK_PIPELINE_STAGE_TRANSFER_BIT, dstStageMask, 0, 0, nullptr, 1,
+	    &barrier, 0, nullptr);
 }
 
 } // namespace filament::backend