+++ filament/backend/src/vulkan/VulkanUtility.h
@@ -56,11 +56,6 @@ VkShaderStageFlags getShaderStageFlags(ShaderStageFlags stageFlags);
 void transitionImageLayout(VkCommandBuffer cmdbuffer, VulkanLayoutTransition transition);
 
 // Helper function for populating barrier fields based on the desired image layout.
-// This logic is specific to blitting.
-VulkanLayoutTransition blitterTransitionHelper(VulkanLayoutTransition transition);
-
-// Helper function for populating barrier fields based on the desired image layout.
-// This logic is specific to texturing.
 VulkanLayoutTransition textureTransitionHelper(VulkanLayoutTransition transition);
 
 bool equivalent(const VkRect2D& a, const VkRect2D& b);